requirements.txt
# AutoAssistGroup Support Ticket System - Dependencies
# Last updated: 2024
# Python 3.9+ required

# Core Flask framework
Flask==2.3.3
Flask-Login>=0.6.2
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database and MongoDB
pymongo==4.5.0
dnspython==2.4.2

# HTTP requests and API integration
requests==2.31.0

# PDF generation and reports
reportlab==4.0.4

# Configuration management
python-dotenv==1.0.0

# Real-time WebSocket support
flask-socketio>=5.3.0
python-socketio>=5.10.0
python-engineio>=4.8.0
eventlet>=0.33.0

# SSL certificates for MongoDB Atlas
certifi>=2023.7.22

# Additional security and performance (optional but recommended)
# Uncomment if using in production:
gunicorn==21.2.0
# redis==5.0.0  # For advanced caching
# celery==5.3.4  # For background tasks

# Timezone support
pytz>=2024.1
